稍微改下接口

+++ Lib for Other/myspi.c
@@ -194,7 +194,7 @@ void myspi_write_byte ( myspi * hspi, uint8_t addr,uint8_t dat )
 	myspi_delay();
 }
 
-void myspi_write_bytes ( myspi * hspi, uint8_t addr, const uint8_t * dat, uint16_t len )
+void myspi_write_bytes ( myspi * hspi, uint8_t addr, const uint8_t * dat, uint32_t len )
 {
 	myspi_transceive_start(hspi);
 	myspi_delay();
@@ -224,7 +224,7 @@ uint8_t myspi_read_byte ( myspi * hspi, uint8_t addr )
 	return dat;
 }
 
-void myspi_read_bytes( myspi * hspi, uint8_t addr, uint8_t * dat, uint16_t len )
+void myspi_read_bytes( myspi * hspi, uint8_t addr, uint8_t * dat, uint32_t len )
 {
 	myspi_transceive_start(hspi);
 	myspi_delay();
@@ -240,7 +240,7 @@ void myspi_read_bytes( myspi * hspi, uint8_t addr, uint8_t * dat, uint16_t len )
 
 /******************* Extend Functions **************************/
 
-void myspi_write ( myspi * hspi, uint8_t * dat, uint16_t len )
+void myspi_write ( myspi * hspi, uint8_t * dat, uint32_t len )
 {
 	myspi_transceive_start(hspi);
 	myspi_delay();
@@ -255,7 +255,7 @@ void myspi_write ( myspi * hspi, uint8_t * dat, uint16_t len )
 	myspi_delay();
 }
 
-void myspi_read ( myspi * hspi, uint8_t * dat, uint16_t len )
+void myspi_read ( myspi * hspi, uint8_t * dat, uint32_t len )
 {
 	myspi_transceive_start(hspi);
 	myspi_delay();
@@ -270,7 +270,7 @@ void myspi_read ( myspi * hspi, uint8_t * dat, uint16_t len )
 	myspi_delay();
 }
 
-void myspi_write_read ( myspi * hspi, uint8_t * dat_in,uint8_t * dat_out, uint16_t len)
+void myspi_write_read ( myspi * hspi, uint8_t * dat_in,uint8_t * dat_out, uint32_t len)
 {
 	myspi_transceive_start(hspi);
 	myspi_delay();
